backend/requirements.txt
envparse
beautifulsoup4==4.9.3
celery==5.0.5
decorator==5.0.7
Django==3.2
gunicorn==20.1.0
imageio==2.9.0
msgpack==1.0.2
numpy==1.20.2
Pillow==8.2.0
psycopg2-binary==2.8.6
redis==3.5.3
requests==2.25.1
scikit-image==0.18.1
django-extensions==3.1.3
django-polymorphic==3.0.0
aiogram==2.13